MSVC builds: Try harder to find msvc_recommended_pragmas.h
We really want to try the best to force-include msvc_recommended_pragmas.h to
check for things at compile time so that we can avoid build breakages due to
stuff like missing includes or attempting to return a value in a function that
is supposed to have a void-return-type, as msvc_recommended_pragmas.h can be
missed out in situations such as when we are building GLib as a fallback
subproject, which happened every now and then as CI couldn't catch things as it
always built GLib as a subproject.

It would also be nice to be quiet from the warnings that we can really
disregard anyways.

So, update the build files to look harder for msvc_recommended_pragmas.h in
GLib's $(includedir), if GLib is not built as a subproject, or otherwise in
the subprojects/glib directory, before just looking for it in the paths in
%INCLUDE%.