theme: Restore flat bg-img on backdrop check/radio
In GTK3, check and radio nodes in the :backdrop state had their backdrop
image removed/flattened, like 'normal' buttons. In GTK4 as of now, this
does not happen. It is because _common.scss simply does not call check()
with the :backdrop state, to generate CSS to flatten the bg-image when
backdropped. This makes that happen, so indicators in backdrop go flat.